On destruction Dispatcher now calls stop to effect and orderly shutdown
Added a missing mutex lock to create_connection()

Finally got around to fixing a missing piece of dispatching to account for the case when there are no watches. Instead, the dispatcher will sleep for the select() timeout period and look to see if there are any new watches.

Removed a hack that flushed every connection at the end of the dispatch loop. The problem was that the sending dispatchers in the examples weren't started.