Fix move constructor for Variant
Move constuctor for Variant was removing the m_signature from both new and other object.
Use std::move for the m_signature member to get the correct behavior.

This was causing wrong message generated for method calls when the Variants were created and passed inline, causing the move constructor to be used.