Add friendly hint when user session not found
In some cases, users may try to connect dbus session as root user
(such as running a program in systemd without the `--user` argument),
which will cause the dbus-cxx library to panic because the control flow
is simply break there, and no caller check if the `m_priv->m_transport` is a
non-null value. These will cause the whole program to panic without any useful
message.

This commit adds pointer validation and add an error message if it fails
to find the `DBUS_SESSION_BUS_ADDRESS` environment variable.

Signed-off-by: Avimitin <dev@avimit.in>