Fixup doxygen (#169)
doc: Remove unused config options

A lot of these are completely unrelated to dbus-cxx (Java and Perl
configuration etc.)

This commit only removes configuration options which are set to their
default value.

This change improves forwards compatibility with newer versions of
Doxygen.

This option is useless and it causes warnings in both old and new
versions of Doxygen.

Newer versions complain that it is not declared as an input file:

warning: Specified markdown mainpage '0-mainpage.md' has not been defined as input file

Having the 0-mainpage.md file both in USE_MDFILE_AS_MAINPAGE and in
INPUT leads to:

/dbus-cxx/build/doc/0-mainpage.md:1: warning: multiple use of section label 'mainpage' while adding anchor, (first occurrence: /dbus-cxx/doc/0-mainpage.md, line 1)

INPUT is set to be recursive, so 0-mainpage.md will be processed. Its
heading also has the #mainpage modifier, so it will be set as the
main page.