doc: Do not copy doc/ to builddir + ignore CWD
The Doxyfile now fully relies on configuration variables which removes
CWD from the equation. This is simpler, more explicit, it removes the
need to copy doc/ to builddir (which wasn't really necessary in the
first place even with the previous system) and it improves Meson support
(Meson prefers either specifying paths explicitly as commandline
arguments or substituting them in .in configuration files).